refactor: split out LocalOfficeContacts
